+++ src/components/Header.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import LevelText from "./LevelText";
 import MeterGroup from "./meters/MeterGroup";
 import XPText from "./XPText";
-import NextLevelIndication from "./NextLevelIndication";
 
 function Header() {
   return (
@@ -10,7 +9,6 @@ function Header() {
       <MeterGroup />
       <XPText />
       <LevelText />
-      <NextLevelIndication />
     </div>
   );
 }