+++ src/components/NextLevelIndication.tsx
@@ -3,9 +3,8 @@ import { FaArrowRight } from "react-icons/fa6";
 
 function NextLevelIndication() {
   return (
-    <div className="flex gap-1 font-extrabold items-center">
-      <FaArrowRight />
-      Next level
+    <div className="fixed top-0 right-0 flex h-full items-center">
+      <FaArrowRight className="text-5xl" />
     </div>
   );
 }